refactored code to break apart components
