+++ components/AdminPage/components/GlobalFilter/index.tsx
@@ -1,0 +1,3 @@
+import GlobalFilter from "./GlobalFilter"
+
+export default GlobalFilter
