+++ components/AdminPage/components/Pagination/index.tsx
@@ -1,0 +1,3 @@
+import Pagination from "./Pagination"
+
+export default Pagination
