+++ components/AdminPage/components/SelectColumFilter/index.tsx
@@ -1,0 +1,3 @@
+import SelectColumnFilter from "./SelectColumnFilter"
+
+export default SelectColumnFilter
