+++ components/AdminPage/components/StatusPill/index.tsx
@@ -1,0 +1,3 @@
+import StatusPill from "./StatusPill"
+
+export default StatusPill
