+++ components/AdminPage/components/TableBody/index.tsx
@@ -1,0 +1,3 @@
+import TableBody from "./TableBody"
+
+export default TableBody
