+++ components/AdminPage/components/TableHeader/index.tsx
@@ -1,0 +1,3 @@
+import TableHeader from "./TableHeader"
+
+export default TableHeader
